сверстано кодом и доделана безопасность
