+++ ImageFeed/Helpers/SceneDelegate.swift
@@ -7,10 +7,17 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
 
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-        // Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
-        // If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
-        // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
-        guard let _ = (scene as? UIWindowScene) else { return }
+        guard let scene = (scene as? UIWindowScene) else { return }
+        //window = UIWindow(windowScene: scene)
+//        window?.rootViewController = UIStoryboard(
+//            name: "Main",
+//            bundle: .main
+//        ).instantiateInitialViewController()
+        let splashViewController = SplashViewController()
+        window = UIWindow(windowScene: scene)
+        window?.rootViewController = splashViewController
+        window?.makeKeyAndVisible()
+
     }
 
     func sceneDidDisconnect(_ scene: UIScene) {
