+++ ImageFeed/OAuth2/OAuth2TokenStorage.swift
@@ -1,0 +1,23 @@
+import Foundation
+import SwiftKeychainWrapper
+
+
+final class OAuth2TokenStorage {
+    var token: String?
+    private static let tokenKey:String = "BearerToken"
+    
+    static var token: String? {
+        get {
+            return KeychainWrapper.standard.string(forKey: tokenKey)
+        }
+        set {
+            if let newValue = newValue {
+                KeychainWrapper.standard.set(newValue, forKey: tokenKey)
+            } else {
+                KeychainWrapper.standard.removeObject(forKey: tokenKey)
+            }
+        }
+    }
+}
+
+
