+++ ImageFeed/ProfileView/ProfileViewController.swift
@@ -1,4 +1,6 @@
 import UIKit
+import Kingfisher
+
 
 class ProfileViewController: UIViewController {
     
@@ -7,26 +9,27 @@ class ProfileViewController: UIViewController {
     private var descriptionLabel: UILabel?
     private let profileService = ProfileService.shared
     private var profileImageServiceObserver: NSObjectProtocol?
+    private var profileImageView: UIImageView? // добаила ссылку на хранение вью
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
         configureProfileImageView()
-                configureUserNameLabel()
+        configureUserNameLabel()
         configureNickNameLabel()
-                configureDescriptionLabel()
-                configureExitButton()
+        configureDescriptionLabel()
+        configureExitButton()
         
-        profileImageServiceObserver = NotificationCenter.default    // 2
-                    .addObserver(
-                        forName: ProfileImageService.DidChangeNotification, // 3
-                        object: nil,                                        // 4
-                        queue: .main                                        // 5
-                    ) { [weak self] _ in
-                        guard let self = self else { return }
-                        self.updateAvatar()                                 // 6
-                    }
-                updateAvatar()
+        profileImageServiceObserver = NotificationCenter.default
+            .addObserver(
+                forName: ProfileImageService.DidChangeNotification,
+                object: nil,
+                queue: .main
+            ) { [weak self] _ in
+                guard let self = self else { return }
+                self.updateAvatar()
+            }
+        updateAvatar()
         
         if let profile = profileService.profile {
             userNameLabel?.text = profile.name
@@ -35,16 +38,18 @@ class ProfileViewController: UIViewController {
         }
     }
     
-    private func updateAvatar() {                                   // 8
-            guard
-                let profileImageURL = ProfileImageService.shared.avatarURL,
-                let url = URL(string: profileImageURL)
-            else { return }
-            // TODO [Sprint 11] Обновить аватар, используя Kingfisher
-        }
+    private func updateAvatar() {
+        guard
+            let profileImageURL = ProfileImageService.shared.avatarURL,
+            let url = URL(string: profileImageURL)
+        else { return }
+        profileImageView?.kf.setImage(with: url)
+        
+    }
     
     
     private func configureProfileImageView() {
+       
         let profileImage = UIImage(named: "Photo")
         let imageView = UIImageView(image: profileImage)
         imageView.tintColor = .ypGray
@@ -54,6 +59,10 @@ class ProfileViewController: UIViewController {
         imageView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 32).isActive = true
         imageView.widthAnchor.constraint(equalToConstant: 70).isActive = true
         imageView.heightAnchor.constraint(equalToConstant: 70).isActive = true
+        profileImageView = imageView
+        imageView.layer.cornerRadius = 35
+        imageView.layer.masksToBounds = true
+
     }
     
     
@@ -68,40 +77,40 @@ class ProfileViewController: UIViewController {
         self.userNameLabel = labelUserName
         
     }
-        
-        
-        private func configureNickNameLabel() {
-            let labelnickName = UILabel()
-            labelnickName.textColor = .ypGray
-            labelnickName.font = UIFont(name: "HelveticaNeue", size: 13)
-            labelnickName.translatesAutoresizingMaskIntoConstraints = false
-            view.addSubview(labelnickName)
-            labelnickName.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20).isActive = true
-            labelnickName.topAnchor.constraint(equalTo: userNameLabel?.bottomAnchor ?? view.safeAreaLayoutGuide.topAnchor, constant: 8).isActive = true
-            self.nickNameLabel = labelnickName
-        }
-        
-        private func configureDescriptionLabel() {
-            let labelDescription = UILabel()
-            labelDescription.textColor = .ypWhite
-            labelDescription.font = UIFont(name: "HelveticaNeue", size: 13)
-            labelDescription.translatesAutoresizingMaskIntoConstraints = false
-            view.addSubview(labelDescription)
-            labelDescription.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20).isActive = true
-            labelDescription.topAnchor.constraint(equalTo: nickNameLabel?.bottomAnchor ?? view.safeAreaLayoutGuide.topAnchor, constant: 8).isActive = true
-            self.descriptionLabel = labelDescription
-        }
-        
-        private func configureExitButton() {
-            let button = UIButton(type: .system)
-            button.setImage(UIImage(named: "Exit"), for: .normal)
-            button.tintColor = .ypRed
-            button.translatesAutoresizingMaskIntoConstraints = false
-            view.addSubview(button)
-            button.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -20).isActive = true
-            button.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 55).isActive = true
-            button.widthAnchor.constraint(equalToConstant: 30).isActive = true
-            button.heightAnchor.constraint(equalToConstant: 30).isActive = true
-            
-        }
+    
+    
+    private func configureNickNameLabel() {
+        let labelnickName = UILabel()
+        labelnickName.textColor = .ypGray
+        labelnickName.font = UIFont(name: "HelveticaNeue", size: 13)
+        labelnickName.translatesAutoresizingMaskIntoConstraints = false
+        view.addSubview(labelnickName)
+        labelnickName.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20).isActive = true
+        labelnickName.topAnchor.constraint(equalTo: userNameLabel?.bottomAnchor ?? view.safeAreaLayoutGuide.topAnchor, constant: 8).isActive = true
+        self.nickNameLabel = labelnickName
     }
+    
+    private func configureDescriptionLabel() {
+        let labelDescription = UILabel()
+        labelDescription.textColor = .ypWhite
+        labelDescription.font = UIFont(name: "HelveticaNeue", size: 13)
+        labelDescription.translatesAutoresizingMaskIntoConstraints = false
+        view.addSubview(labelDescription)
+        labelDescription.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20).isActive = true
+        labelDescription.topAnchor.constraint(equalTo: nickNameLabel?.bottomAnchor ?? view.safeAreaLayoutGuide.topAnchor, constant: 8).isActive = true
+        self.descriptionLabel = labelDescription
+    }
+    
+    private func configureExitButton() {
+        let button = UIButton(type: .system)
+        button.setImage(UIImage(named: "Exit"), for: .normal)
+        button.tintColor = .ypRed
+        button.translatesAutoresizingMaskIntoConstraints = false
+        view.addSubview(button)
+        button.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -20).isActive = true
+        button.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 55).isActive = true
+        button.widthAnchor.constraint(equalToConstant: 30).isActive = true
+        button.heightAnchor.constraint(equalToConstant: 30).isActive = true
+        
+    }
+}
