+++ ImageFeed/Resourses/Assets.xcassets/tab_profile_active.imageset/Contents.json
