+++ ImageFeed/Resourses/Assets.xcassets/tab_profile_active.imageset/tub_ profile_active.png
