+++ ImageFeed/SingleImage/SingleImageViewController.swift
@@ -31,7 +31,7 @@ class SingleImageViewController: UIViewController {
         rescaleAndCenterImageInScrollView(image: image)
         scrollView.minimumZoomScale = 0.1
         scrollView.maximumZoomScale = 1.25
-
+        
     }
     
     private func rescaleAndCenterImageInScrollView(image: UIImage) {
