+++ ImageFeed/SplashView/SplashViewController.swift
@@ -3,25 +3,36 @@ import ProgressHUD
 
 final class SplashViewController: UIViewController {
     
-    
     private let ShowAuthenticationScreenSegueIdentifier = "ShowAuthenticationScreen"
     private let oauth2Service = OAuth2Service()
     private let oauth2TokenStorage = OAuth2TokenStorage()
-    //var profileService: ProfileService?
     private let profileService = ProfileService.shared
     
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         
-
+        let imageView = UIImageView(image: UIImage(named: "auth_screen_logo"))
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        view.addSubview(imageView)
+        
+        NSLayoutConstraint.activate([
+            imageView.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            imageView.centerYAnchor.constraint(equalTo: view.centerYAnchor)
+        ])
+        
         if let token = OAuth2TokenStorage.token {
-              UIBlockingProgressHUD.show()
-              fetchProfile(token: token)
+            UIBlockingProgressHUD.show()
+            fetchProfile(token: token)
             switchToTabBarController()
-          } else {
-              performSegue(withIdentifier: ShowAuthenticationScreenSegueIdentifier, sender: nil)
-          }
-      }
+        } else {
+            let storyboard = UIStoryboard(name: "Main", bundle: nil)
+            if let authViewController = storyboard.instantiateViewController(withIdentifier: "AuthViewController") as? AuthViewController {
+                authViewController.modalPresentationStyle = .fullScreen
+                authViewController.delegate = self
+                present(authViewController, animated: true, completion: nil)
+            }
+        }
+    }
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
@@ -49,30 +60,15 @@ final class SplashViewController: UIViewController {
                 print("Network request failed: \(error)")
                 UIBlockingProgressHUD.dismiss()
                 DispatchQueue.main.async {
-                       let alertController = UIAlertController(title: "Что-то пошло не так(", message: "Не удалось загрузить профиль", preferredStyle: .alert)
-                       alertController.addAction(UIAlertAction(title: "Ок", style: .default))
-                       self.present(alertController, animated: true, completion: nil)
-                   }
+                    let alertController = UIAlertController(title: "Что-то пошло не так(", message: "Не удалось загрузить профиль", preferredStyle: .alert)
+                    alertController.addAction(UIAlertAction(title: "Ок", style: .default))
+                    self.present(alertController, animated: true, completion: nil)
+                }
                 break
             }
         }
     }
 }
-
-extension SplashViewController {
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        if segue.identifier == ShowAuthenticationScreenSegueIdentifier {
-            guard
-                let navigationController = segue.destination as? UINavigationController,
-                let viewController = navigationController.viewControllers[0] as? AuthViewController
-            else { fatalError("Failed to prepare for \(ShowAuthenticationScreenSegueIdentifier)") }
-            viewController.delegate = self
-        } else {
-            super.prepare(for: segue, sender: sender)
-        }
-    }
-}
-
 
 extension SplashViewController: AuthViewControllerDelegate {
     func authViewController(_ vc: AuthViewController, didAuthenticateWithCode code: String) {
@@ -83,13 +79,10 @@ extension SplashViewController: AuthViewControllerDelegate {
         }
     }
     
-
-    
     private func fetchOAuthToken(_ code: String) {
         
-        //oauth2Service.fetchOAuthToken(code) { [weak self] result in
         OAuth2Service.shared.fetchOAuthToken(code) { [weak self] result in
-
+            
             guard let self = self else { return }
             switch result {
             case .success(let authToken):
@@ -102,10 +95,10 @@ extension SplashViewController: AuthViewControllerDelegate {
             case .failure:
                 UIBlockingProgressHUD.dismiss()
                 DispatchQueue.main.async {
-                       let alertController = UIAlertController(title: "Что-то пошло не так(", message: "Не удалось войти в систему", preferredStyle: .alert)
-                       alertController.addAction(UIAlertAction(title: "Ок", style: .default))
-                       self.present(alertController, animated: true, completion: nil)
-                   }
+                    let alertController = UIAlertController(title: "Что-то пошло не так(", message: "Не удалось войти в систему", preferredStyle: .alert)
+                    alertController.addAction(UIAlertAction(title: "Ок", style: .default))
+                    self.present(alertController, animated: true, completion: nil)
+                }
                 break
             }
         }
